# setup/e2mgr/Swagger/E2M_API_2020-04-23.yaml
openapi: 3.0.0
info:
  title: E2 Manager Service
  description: E2 Manager Service APIs
  version: 4.4.4
servers:
  - url: 'http://{apiRoot}/v1'
    variables:
      apiRoot:
        default: 'localhost:3800'
paths:
...
  '/nodeb/{ranName}/update':
    put:
      summary: Update GNB
      tags:
        - nodeb
      operationId: UpdateGnb
      parameters:
        - name: ranName
          in: path
          required: true
          description: Name of GNB RAN to update
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGnbRequest'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateGnbResponse'
        '400':
          description: Invalid input
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  '/nodeb/ids':
    get:
      tags:
        - nodeb
      summary: Get RANs identities list
      operationId: getNodebIdList
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NodebIdentity'
        '500':
          description: Internal Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    UpdateGnbRequest:
      type: object
      required:
        - servedNrCells
      properties:
            servedNrCells:
              items:
                required:
                  - servedNrCellInformation
                properties:
                  nrNeighbourInfos:
                    items:
                      required:
                        - choiceNrMode
                        - nrCgi 
                        - nrMode
                        - nrPci
                      properties:
                        choiceNrMode:
                          properties:
                            fdd:
                              properties:
                                dlarFcnFreqInfo:
                                  properties:
                                    frequencyBands:
                                      items:
                                        properties:
                                          nrFrequencyBand:
                                            type: integer
                                          supportedSulBands:
                                            items:
                                              type: integer
                                            type: array
                                        additionalProperties: false
                                        type: object
                                      type: array
                                    nrArFcn:
                                      oneOf:
                                        - type: integer
                                    sulInformation:
                                      properties:
                                        sulArFcn:
                                          oneOf:
                                            - type: integer
                                        sulTransmissionBandwidth:
                                          properties:
                                            ncnrb:
                                              oneOf:
                                                - type: string
                                                - type: integer
                                            nrscs:
                                              oneOf:
                                                - type: string
                                                - type: integer
                                          additionalProperties: false
                                          type: object
                                      additionalProperties: false
                                      type: object
                                  additionalProperties: false
                                  type: object
                                ularFcnFreqInfo:
                                  properties:
                                    frequencyBands:
                                      items:
                                        properties:
                                          nrFrequencyBand:
                                            type: integer
                                          supportedSulBands:
                                            items:
                                              type: integer
                                            type: array
                                        additionalProperties: false
                                        type: object
                                      type: array
                                    nrArFcn:
                                      oneOf:
                                        - type: integer
                                    sulInformation:
                                      properties:
                                        sulArFcn:
                                          oneOf:
                                            - type: integer
                                        sulTransmissionBandwidth:
                                          properties:
                                            ncnrb:
                                              oneOf:
                                                - type: string
                                                - type: integer
                                            nrscs:
                                              oneOf:
                                                - type: string
                                                - type: integer
                                          additionalProperties: false
                                          type: object
                                      additionalProperties: false
                                      type: object
                                  additionalProperties: false
                                  type: object
                              additionalProperties: false
                              type: object
                            tdd:
                              properties:
                                arFcnNrFreqInfo:
                                  properties:
                                    frequencyBands:
                                      items:
                                        properties:
                                          nrFrequencyBand:
                                            type: integer
                                          supportedSulBands:
                                            items:
                                              type: integer
                                            type: array
                                        additionalProperties: false
                                        type: object
                                      type: array
                                    nrArFcn:
                                      oneOf:
                                        - type: integer
                                    sulInformation:
                                      properties:
                                        sulArFcn:
                                          oneOf:
                                            - type: integer
                                        sulTransmissionBandwidth:
                                          properties:
                                            ncnrb:
                                              oneOf:
                                                - type: string
                                                - type: integer
                                            nrscs:
                                              oneOf:
                                                - type: string
                                                - type: integer
                                          additionalProperties: false
                                          type: object
                                      additionalProperties: false
                                      type: object
                                  additionalProperties: false
                                  type: object
                              additionalProperties: false
                              type: object
                          additionalProperties: false
                          type: object
                        configuredStac:
                          type: string
                        nrCgi:
                          type: string
                        nrMode:
                          oneOf:
                            - type: string
                            - type: integer
                        nrPci:
                          type: integer
                        stac5g:
                          type: string
                      additionalProperties: false
                      type: object
                    type: array
                  servedNrCellInformation:
                    required:
                      - cellId
                      - choiceNrMode
                      - nrMode
                      - nrPci
                      - servedPlmns
                    properties:
                      cellId:
                        type: string
                      choiceNrMode:
                        properties:
                          fdd:
                            properties:
                              dlFreqInfo:
                                properties:
                                  frequencyBands:
                                    items:
                                      properties:
                                        nrFrequencyBand:
                                          type: integer
                                        supportedSulBands:
                                          items:
                                            type: integer
                                          type: array
                                      additionalProperties: false
                                      type: object
                                    type: array
                                  nrArFcn:
                                    oneOf:
                                      - type: integer
                                  sulInformation:
                                    properties:
                                      sulArFcn:
                                        oneOf:
                                          - type: integer
                                      sulTransmissionBandwidth:
                                        properties:
                                          ncnrb:
                                            oneOf:
                                              - type: string
                                              - type: integer
                                          nrscs:
                                            oneOf:
                                              - type: string
                                              - type: integer
                                        additionalProperties: false
                                        type: object
                                    additionalProperties: false
                                    type: object
                                additionalProperties: false
                                type: object
                              dlTransmissionBandwidth:
                                properties:
                                  ncnrb:
                                    oneOf:
                                      - type: string
                                      - type: integer
                                  nrscs:
                                    oneOf:
                                      - type: string
                                      - type: integer
                                additionalProperties: false
                                type: object
                              ulFreqInfo:
                                properties:
                                  frequencyBands:
                                    items:
                                      properties:
                                        nrFrequencyBand:
                                          type: integer
                                        supportedSulBands:
                                          items:
                                            type: integer
                                          type: array
                                      additionalProperties: false
                                      type: object
                                    type: array
                                  nrArFcn:
                                    oneOf:
                                      - type: integer
                                  sulInformation:
                                    properties:
                                      sulArFcn:
                                        oneOf:
                                          - type: integer
                                      sulTransmissionBandwidth:
                                        properties:
                                          ncnrb:
                                            oneOf:
                                              - type: string
                                              - type: integer
                                          nrscs:
                                            oneOf:
                                              - type: string
                                              - type: integer
                                        additionalProperties: false
                                        type: object
                                    additionalProperties: false
                                    type: object
                                additionalProperties: false
                                type: object
                              ulTransmissionBandwidth:
                                properties:
                                  ncnrb:
                                    oneOf:
                                      - type: string
                                      - type: integer
                                  nrscs:
                                    oneOf:
                                      - type: string
                                      - type: integer
                                additionalProperties: false
                                type: object
                            additionalProperties: false
                            type: object
                          tdd:
                            properties:
                              nrFreqInfo:
                                properties:
                                  frequencyBands:
                                    items:
                                      properties:
                                        nrFrequencyBand:
                                          type: integer
                                        supportedSulBands:
                                          items:
                                            type: integer
                                          type: array
                                      additionalProperties: false
                                      type: object
                                    type: array
                                  nrArFcn:
                                    oneOf:
                                      - type: integer
                                  sulInformation:
                                    properties:
                                      sulArFcn:
                                        oneOf:
                                          - type: integer
                                      sulTransmissionBandwidth:
                                        properties:
                                          ncnrb:
                                            oneOf:
                                              - type: string
                                              - type: integer
                                          nrscs:
                                            oneOf:
                                              - type: string
                                              - type: integer
                                        additionalProperties: false
                                        type: object
                                    additionalProperties: false
                                    type: object
                                additionalProperties: false
                                type: object
                              transmissionBandwidth:
                                properties:
                                  ncnrb:
                                    oneOf:
                                      - type: string
                                      - type: integer
                                  nrscs:
                                    oneOf:
                                      - type: string
                                      - type: integer
                                additionalProperties: false
                                type: object
                            additionalProperties: false
                            type: object
                        additionalProperties: false
                        type: object
                      configuredStac:
                        type: string
                      nrMode:
                        oneOf:
                          - type: string
                          - type: integer
                      nrPci:
                        type: integer
                      servedPlmns:
                        items:
                          type: string
                        type: array
                      stac5g:
                        type: string
                    additionalProperties: false
                    type: object
                additionalProperties: false
                type: object
              type: array
      additionalProperties: false
...
    NodebIdentity:
      properties:
        globalNbId:
          properties:
            nbId:
              type: string
            plmnId:
              type: string
          type: object
        inventoryName:
          type: string
      type: object
...
    ErrorResponse:
      type: object
      required:
        - errorCode
        - errorMessage
      properties:
        errorCode:
          type: string
          description: >-
            401 - corrupted json, 402 - validation error, 403 - RAN in wrong
            state, 404 - resource not found, 500 - RNIB error, 501 - internal
            problem, 502 - RMR error, 503 - Routing Manager Unavailable
        errorMessage:
          type: string
          description: Human readable text
...